requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/base.in
#
amqp==5.0.6
    # via kombu
apscheduler==3.6.3
    # via python-telegram-bot
asgiref==3.4.1
    # via django
billiard==3.6.4.0
    # via celery
cachetools==4.2.2
    # via python-telegram-bot
celery==5.1.2
    # via
    #   -r requirements/base.in
    #   django-celery-beat
    #   django-celery-results
    #   flower
certifi==2021.5.30
    # via
    #   python-telegram-bot
    #   requests
charset-normalizer==2.0.4
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
dateparser==1.0.0
    # via -r requirements/base.in
django==3.2.6
    # via
    #   -r requirements/base.in
    #   django-celery-beat
    #   django-timezone-field
django-celery-beat==2.2.1
    # via -r requirements/base.in
django-celery-results==2.2.0
    # via -r requirements/base.in
django-timezone-field==4.2.1
    # via django-celery-beat
feedparser==6.0.8
    # via -r requirements/base.in
flower==1.0.0
    # via -r requirements/base.in
gunicorn==20.1.0
    # via -r requirements/base.in
humanize==3.11.0
    # via flower
idna==3.2
    # via requests
kombu==5.1.0
    # via celery
prometheus-client==0.11.0
    # via flower
prompt-toolkit==3.0.19
    # via click-repl
psycopg2-binary==2.9.1
    # via -r requirements/base.in
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   dateparser
    #   python-crontab
python-telegram-bot==13.7
    # via -r requirements/base.in
pytz==2021.1
    # via
    #   apscheduler
    #   celery
    #   dateparser
    #   django
    #   django-timezone-field
    #   flower
    #   python-telegram-bot
    #   tzlocal
redis==3.5.3
    # via -r requirements/base.in
regex==2021.8.3
    # via dateparser
requests==2.26.0
    # via -r requirements/base.in
ruamel.yaml==0.17.10
    # via -r requirements/base.in
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   apscheduler
    #   click-repl
    #   python-dateutil
sqlparse==0.4.1
    # via django
tornado==6.1
    # via
    #   flower
    #   python-telegram-bot
tzlocal==2.1
    # via
    #   apscheduler
    #   dateparser
urllib3==1.26.6
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
